authapi/requirements.txt
aiohttp==3.9.3 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
alembic==1.13.1 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
async-fastapi-jwt-auth==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.12.0"
asyncpg==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==42.0.3 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.14 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi-limiter==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
fastapi-oauth2==1.0.0b3 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.100.1 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.3 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.26.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
mako==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.22.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-jaeger-thrift==1.21.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.43b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.43b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.43b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.22.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.43b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.43b0 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.9.14 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.16.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-asyncio==0.23.5 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.4.4 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
redis==5.0.1 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==69.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-core==4.5.3 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.27 ; python_version >= "3.10" and python_version < "4.0"
sqlmodel==0.0.14 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
thrift==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
ua-parser==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
user-agents==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.26.0 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
sentry-sdk[fastapi]
